Definindo services e repositórios os gamification

+++ Gamification/src/main/kotlin/com/pitzdev/gamification/models/leaderBord/LeaderBoard.kt
@@ -4,4 +4,4 @@ import com.pitzdev.gamification.models.base.BaseEntity
 import javax.persistence.Entity
 
 @Entity
-data class LeaderBoard(var userId: Long, var totalScore: Int) : BaseEntity()
+data class LeaderBoard(var userId: Long, var totalScore: Long) : BaseEntity()
